Fixed user-/group-info error raising when user/group does not exist
POSIX only sets errno for things like running out of file
descriptors (these are generally library calls that look at files), if
the user or group does not exist, it's not considered an error.  So we
must call srfi-170-error.

An example of the general problem that the tests don't check to see if
valid SRFI 170 objects are raised.